Moves models that don't inherit Status to their own files
